requirements.txt
transformers==4.37.2
torch==2.2.0
Pillow==10.2.0
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.6
numpy==1.26.3
supabase==1.2.0
python-dotenv==1.0.0 